requirements.txt
fastapi==0.88.0
pymongo==4.3.3
beanie==1.16.8
uvicorn==0.20.0
pydantic[dotenv]
pandas==1.5.2
yfinance==0.2.3
motor==3.1.1 

python-dateutil==2.8.2
pytest==7.2.1